Variational Embeddings of Animal Motion/requirements.txt
tensorflow
pyvis
networkx
plotly
umap-learn
scikit-learn
pandas
scipy>=1.7,<2.0
typing-extensions==4.1.1